+++ packages/babel-jest/src/index.js
@@ -6,15 +6,14 @@
  * of patent rights can be found in the PATENTS file in the same directory.
  */
 
+'use strict';
+
 const babel = require('babel-core');
 const jestPreset = require('babel-preset-jest');
-const path = require('path');
-
-const NODE_MODULES = path.sep + 'node_modules' + path.sep;
 
 module.exports = {
   process(src, filename) {
-    if (!filename.includes(NODE_MODULES) && babel.util.canCompile(filename)) {
+    if (babel.util.canCompile(filename)) {
       return babel.transform(src, {
         filename,
         presets: [jestPreset],