+++ src/HasteModuleLoader/__tests__/HasteModuleLoader-jsdom-env-test.js
@@ -11,6 +11,8 @@
 
 jest.autoMockOff();
 
+jasmine.DEFAULT_TIMEOUT_INTERVAL = 15000;
+
 const path = require('path');
 const utils = require('../../lib/utils');